Exif DateTimeOriginal: don't parse using std::regex
Parse the "2019:11:20 23:07:39" string in a primitive way and
don't care if it's not really a date and time. If it's not then
it's broken on purpose, but that has no security implications for
olymp -- we just print the data we find.

I used to parse it using std::regex, but I discovered that this
bloated the 25 kB olymp binary to 105 kB and I didn't feel that
was appropriate.