print some timestamps as hh:mm:ss
Sometimes it's useful to know the time the photo was taken
with better precision than a minute.  Clusters of photos are
printed with seconds in the time stamp, if they were taken
with gaps of no more than 60 seconds.

This is useful because if you, for example, took two photos of
a plant within fifteen seconds, it's probably the same individual
in both of them.

This feature is enabled by default, but the -M option disables it.